requirements.txt
python-decouple==3.4
Flask==1.1.2
Flask-RESTful==0.3.8
PyJWT==2.0.1
bcrypt==3.2.0
pymongo==3.11.3